pyproject.toml
[tool.poetry]
name = "docker-deploy"
version = "0.1.0"
description = ""
authors = ["MaxAFriedrich <58374387+MaxAFriedrich@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "docker_deploy"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
ansible = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
